test: Add test PDF file for Marker integration tests
- Include test_file.pdf with 2 pages for PDF processing tests
