+++ tests/resources/test_file.pdf
@@ -1,0 +1,63 @@
+%PDF-1.4
+1 0 obj
+<< /Type /Catalog /Pages 2 0 R >>
+endobj
+2 0 obj
+<< /Type /Pages /Kids [3 0 R 4 0 R] /Count 2 >>
+endobj
+3 0 obj
+<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792]
+   /Contents 5 0 R /Resources << /Font << /F1 6 0 R >> >> >>
+endobj
+4 0 obj
+<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792]
+   /Contents 7 0 R /Resources << /Font << /F1 6 0 R >> >> >>
+endobj
+5 0 obj
+<< /Length 178 >>
+stream
+BT
+/F1 24 Tf
+100 700 Td
+(OpenRAG Test Document) Tj
+0 -40 Td
+/F1 12 Tf
+(This is a test PDF file for integration testing.) Tj
+0 -20 Td
+(It contains multiple pages to test Marker processing.) Tj
+ET
+endstream
+endobj
+6 0 obj
+<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica >>
+endobj
+7 0 obj
+<< /Length 156 >>
+stream
+BT
+/F1 18 Tf
+100 700 Td
+(Page 2: Additional Content) Tj
+0 -30 Td
+/F1 12 Tf
+(This is the second page of the test document.) Tj
+0 -20 Td
+(Marker should correctly identify page boundaries.) Tj
+ET
+endstream
+endobj
+xref
+0 8
+0000000000 65535 f
+0000000009 00000 n
+0000000058 00000 n
+0000000115 00000 n
+0000000242 00000 n
+0000000369 00000 n
+0000000599 00000 n
+0000000676 00000 n
+trailer
+<< /Size 8 /Root 1 0 R >>
+startxref
+884
+%%EOF
